Conflicts:
	ServerWithConfiguration/src/main/java/info/smart_tools/smartactors/sandbox/server_with_configuration/ServerWithConfiguration.java
	pom.xml

+++ CreateAsyncOperation/src/main/java/info/smart_tools/smartactors/actors/create_async_operation/wrapper/CreateAsyncOperationMessage.java
@@ -21,7 +21,7 @@ public interface CreateAsyncOperationMessage {
      * @return TTL for async operation
      * @throws ReadValueException if error during get is occurred
      */
-    Long getExpiredTime() throws ReadValueException;
+    Integer getExpiredTime() throws ReadValueException;
 
     /**
      * Setter
